Update index.md
I edited this text and made my changes to this branch. One remaining question I had was about the output directory option, which states "If not specified, the BDBag will be created in-place in DATA_PATH if necessary." What does "in-place" mean? Will the BDBag replace DATA PATH? Or be placed within it?